requirements.txt
PyQt6
reportlab
python-dotenv
PyMuPDF
google-generativeai